+++ src/manager.rs
