--- src/new/mod.rs
@@ -1,15 +1,0 @@
-mod manager;
-mod options;
-mod parameters;
-mod pool;
-mod runtime;
-mod worker;
-
-pub use manager::{ModuleDefinition, WorkerManager};
-pub use options::{
-    DomainPermission, ExecOptions, ModuleOptions, ResolvedExecOptions, WorkerOptions,
-};
-pub use parameters::{FunctionParameters, TryIntoFunctionParameters};
-pub use pool::{Pool, PoolBuilder};
-pub use runtime::{Runtime, RuntimeError};
-pub use worker::{Worker, WorkerBuilder};
